Connected functionality with all the service classes to GUI. (Database -> GUI) And cleaned out some code.

+++ WardrobeApp/src/model/dto/Style.java
@@ -1,26 +1,29 @@
 package model.dto;
 
 public class Style {
-    private Long id;
+    private int id;
     private String name;
-    public Style(Long id, String name) {
+
+    public Style(int id, String name) {
         this.id = id;
         this.name = name;
     }
+
     public Style() {
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
     public String getName() {
         return name;
     }
+
     public void setName(String name) {
         this.name = name;
     }
